folder structure -removed staticLevel folder and added data too src/data. -moved levels, legends to their own js files inside of src/data. -renamed index.js to dinos.js and left the dinos variable in the file. -adjusted imports on level index.js and legend index.js
