+++ src/components/data/legends.js
@@ -1,0 +1,84 @@
+export const legends = {
+    //red - 8 different legend boards
+    '1': {
+      '1': 1,
+      '2': 2,
+      '3': 3
+    },
+    '2': {
+      '2': 2,
+      '3': 3,
+      '4': 5
+    },
+    '3': {
+      '3': 3,
+      '4': 5,
+      '5': 10
+    },
+    '4': {
+      '1': 1,
+      '2': 2,
+      '5': 10
+    },
+    '5': {
+      '1': 1,
+      '2': 2,
+      '4': 5
+    },
+    '6': {
+      '1': 1,
+      '3': 3,
+      '5': 10
+    },
+    '7': {
+      '1': 1,
+      '4': 5,
+      '5': 10
+    },
+    '8': {
+      '2': 2,
+      '3': 3,
+      '5': 10
+    },
+    //blue - 8 different legend boards
+    '9': {
+      '6': 1,
+      '7': 2,
+      '8': 3
+    },
+    '10': {
+      '7': 2,
+      '8': 3,
+      '9': 5
+    },
+    '11': {
+      '8': 3,
+      '9': 5,
+      '10': 10
+    },
+    '12': {
+      '6': 1,
+      '7': 2,
+      '10': 10
+    },
+    '13': {
+      '6': 1,
+      '7': 2,
+      '9': 5
+    },
+    '14': {
+      '6': 1,
+      '8': 3,
+      '10': 10
+    },
+    '15': {
+      '6': 1,
+      '9': 5,
+      '10': 10
+    },
+    '16': {
+      '7': 2,
+      '9': 5,
+      '10': 10
+    },
+  }
