+++ src/components/data/levels.js
@@ -1,0 +1,103 @@
+export const levels = { //first three levels are fixed winners.
+    '1': {
+      // range of number of dinos to be generated
+      range: {
+        low: 1,
+        high: 1,
+      },
+      legends: {
+        red: 5,
+        blue: 11,
+      }
+    },
+    '2': {
+      range: {
+        low: 1,
+        high: 3,
+      },
+      legends: {
+        red: 4,
+        blue: 9,
+      }
+    },
+    '3': {
+      range: {
+        low: 2,
+        high: 2,
+      },
+      legends: {
+        red: 3,
+        blue: 14,
+      }
+    },
+    '4': {
+      range: {
+        low: 1,
+        high: 4,
+      },
+      legends: {
+        red: 1,
+        blue: 10,
+      }
+    },
+    '5': {
+      range: {
+        low: 1,
+        high: 4,
+      },
+      legends: {
+        red: 8,
+        blue: 11,
+      }
+    },
+    '6': {
+      range: {
+        low: 1,
+        high: 4,
+      },
+      legends: {
+        red: 7,
+        blue: 12,
+      }
+    },
+    '7': {
+      range: {
+        low: 1,
+        high: 4,
+      },
+      legends: {
+        red: 6,
+        blue: 13,
+      }
+    },
+    '8': {
+      range: {
+        low: 1,
+        high: 4,
+      },
+      legends: {
+        red: 5,
+        blue: 14,
+      }
+    },
+    '9': {
+      range: {
+        low: 1,
+        high: 4,
+      },
+      legends: {
+        red: 4,
+        blue: 15,
+      }
+    },
+    '10': {
+      range: {
+        low: 1,
+        high: 4,
+      },
+      legends: {
+        red: 3,
+        blue: 16,
+      }
+    },
+  }
